requirements-xbee.txt
pyserial
numpy
pandas
requests
